% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{FoodDetails-class}
\alias{FoodDetails-class}
\title{FoodDetails}
\description{
FoodDetails
}
\section{Slots}{

\describe{
\item{\code{fdc_id}}{A food id whose details are required.}
}}

