% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{FoodSearch-class}
\alias{FoodSearch-class}
\title{FoodSearch Class}
\description{
FoodSearch Class
}
\section{Slots}{

\describe{
\item{\code{search_phrase}}{A phrase to use when querying the database}
}}

