% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodDefinitions.R
\docType{methods}
\name{get_food_info}
\alias{get_food_info}
\alias{get_food_info,FoodSearch-method}
\title{Method definitions for the FoodSearch Class}
\usage{
get_food_info(
  object,
  target = "description",
  page_number = 1,
  sort_field = "publishedDate",
  sort_direction = "asc",
  brand_owner = NULL,
  ...
)

\S4method{get_food_info}{FoodSearch}(
  object,
  target = "description",
  page_number = 1,
  sort_field = "publishedDate",
  sort_direction = "asc",
  brand_owner = NULL
)
}
\arguments{
\item{object}{An object of class FoodSearch}

\item{target}{Target field to extract, can be a vector. Defaults to "description"}

\item{page_number}{Page number of results to return. Defaults to 1}

\item{sort_field}{Character specifying what field to use for sorting. 
Defaults to publicationDate}

\item{sort_direction}{One of 'asc' or 'desc' to sort results in ascending or 
descending order respectively.}

\item{brand_owner}{Brand owner.}

\item{...}{Ignored.}
}
\description{
Method definitions for the FoodSearch Class
}
\examples{
\dontrun{
my_object <- make_object("FoodSearch",search_phrase="indomie")
get_food_info(my_object,"description")
}
}
