% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{key_signup}
\alias{key_signup}
\title{Directs user to the api key signup page}
\usage{
key_signup()
}
\value{
Opens a browser tab to the FDC api key sign up page
}
\description{
Directs user to the api key signup page
}
