% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_object}
\alias{make_object}
\title{Makes an object of either FoodDetails or FoodSearch}
\usage{
make_object(object_class = NULL, ...)
}
\arguments{
\item{object_class}{Class of the object eg FoodSearch}

\item{...}{Other arguments to the object's class definitions.}
}
\description{
Makes an object of either FoodDetails or FoodSearch
}
