% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_api_key}
\alias{set_api_key}
\title{Sets a user's session API key}
\usage{
set_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{An api key as obtained from Food Data Central}
}
\value{
Sets the session api key
}
\description{
Sets a user's session API key
}
\examples{
\dontrun{
set-api_key("my key here")
}
}
