% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{FoodSearch-class}
\alias{FoodSearch-class}
\title{FoodSearch Class}
\description{
FoodSearch Class
}
\section{Slots}{

\describe{
\item{\code{api_key}}{An api key as obtained from FoodDataCentral}

\item{\code{search_phrase}}{A phrase to use when querying the database}

\item{\code{brand_owner}}{A character specifying who owns the brand e.g usa}

\item{\code{ingredients}}{An optional ingredients phrase}
}}

