% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodDefinitions.R
\name{get_food_details}
\alias{get_food_details}
\title{Method definitions for the FoodDetails Class}
\usage{
get_food_details(object, ...)
}
\arguments{
\item{object}{An object of class FoodSearch}

\item{target-field}{Target field to extract, can be a vector. Defaults to "description"}
}
\description{
Method definitions for the FoodDetails Class
}
\examples{
\dontrun{
test_object <-make_object("FoodDetails",fdc_id = 504905)
get_food_details(test_object, "ingredients")
}
}
