% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodDefinitions.R
\name{get_nutrients}
\alias{get_nutrients}
\title{Get Nutrients given a FoodDetails Object}
\usage{
get_nutrients(object, ...)
}
\arguments{
\item{object}{An object of class FoodSearch}
}
\value{
A `data.frame` object showing nutrients of a given food based on FDC ID.
}
\description{
Get Nutrients given a FoodDetails Object
}
\examples{
\dontrun{
test_object <-make_object("FoodDetails",fdc_id = 504905)
head(get_nutrients(test_object))
}
}
